// src/BindingChaos.Web/src/api/models/CommonsListItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BindingChaos.Web.Gateway
 * BindingChaos Web Gateway API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Represents a commons item in a list response.
 * @export
 * @interface CommonsListItemResponse
 */
export interface CommonsListItemResponse {
    /**
     * The unique identifier of the commons.
     * @type {string}
     * @memberof CommonsListItemResponse
     */
    id?: string;
    /**
     * The name of the commons.
     * @type {string}
     * @memberof CommonsListItemResponse
     */
    name?: string;
    /**
     * The description of the commons.
     * @type {string}
     * @memberof CommonsListItemResponse
     */
    description?: string;
    /**
     * The lifecycle status of the commons.
     * @type {string}
     * @memberof CommonsListItemResponse
     */
    status?: string;
    /**
     * The pseudonym of the participant who proposed the commons.
     * @type {string}
     * @memberof CommonsListItemResponse
     */
    proposedByPseudonym?: string;
    /**
     * The timestamp when the commons was proposed.
     * @type {Date}
     * @memberof CommonsListItemResponse
     */
    proposedAt?: Date;
}

/**
 * Check if a given object implements the CommonsListItemResponse interface.
 */
export function instanceOfCommonsListItemResponse(value: object): value is CommonsListItemResponse {
    return true;
}

export function CommonsListItemResponseFromJSON(json: any): CommonsListItemResponse {
    return CommonsListItemResponseFromJSONTyped(json, false);
}

export function CommonsListItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CommonsListItemResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'status': json['status'] == null ? undefined : json['status'],
        'proposedByPseudonym': json['proposedByPseudonym'] == null ? undefined : json['proposedByPseudonym'],
        'proposedAt': json['proposedAt'] == null ? undefined : (new Date(json['proposedAt'])),
    };
}

export function CommonsListItemResponseToJSON(json: any): CommonsListItemResponse {
    return CommonsListItemResponseToJSONTyped(json, false);
}

export function CommonsListItemResponseToJSONTyped(value?: CommonsListItemResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'status': value['status'],
        'proposedByPseudonym': value['proposedByPseudonym'],
        'proposedAt': value['proposedAt'] == null ? undefined : ((value['proposedAt']).toISOString()),
    };
}